CLUE-568: relocate rete files out of nodes/; fix collapsed-group connect/render bugs
Story 5 (file relocation, Teale review):
- move rete-manager.tsx (+ its 3 test files) from nodes/ to a new rete/ dir
- move marquee-selection.ts (+ test) from nodes/ to utilities/
- rewrite internal + external imports accordingly (nodes/ siblings -> ../nodes/)

Bug fixes (QA):
- collapsed-group proxy sockets are keyboard-only (remove the mouse onClick that
  could enter connect mode and leave candidate-highlight rings stuck until reload)
- cancel an in-progress keyboard connection on any outside pointerdown, so the
  highlight rings never persist (also fixes the same gap in CLUE-455 connect mode)
- re-apply collapsed-connection visibility when a connection view renders, so a
  connection into a collapsed group hides correctly on load, not only after an
  interaction (its view renders async, after the fireImmediately reaction pass)
