+++ src/plugins/dataflow/components/dataflow-rete-manager-context.tsx
@@ -1,4 +1,4 @@
 import React from "react";
-import { ReteManager } from "../nodes/rete-manager";
+import { ReteManager } from "../rete/rete-manager";
 
 export const DataflowReteManagerContext = React.createContext<ReteManager | null>(null);
