+++ src/plugins/dataflow/components/ui/dataflow-groups-overlay.scss
@@ -135,7 +135,6 @@
       border-radius: 50%;
       background: #949494;
       transform: translate(-50%, -50%);
-      cursor: pointer;
 
       &.input { left: 0; }
       &.output { left: 100%; }
