+++ src/plugins/dataflow/rete/connection-mode.test.ts
