+++ src/plugins/dataflow/utilities/marquee-selection.ts
